Rework time and timezone handling
- Upgrade to Python 3.9
This is needed because of some strange behavior observed from using
`pytz.timezone`:
```
>>> from datetime import datetime, timedelta
>>> from pytz import timezone
>>> datetime(2020, 10, 7, tzinfo=timezone("America/Los_Angeles")) \
  .isoformat()
'2020-10-07T00:00:00-07:53'
#                    ^^^^^
#                    wut?
```
This was literally the correct time offset for LA in, like 1890. But why
does pytz think this is correct given the date provided? Time is hard.

Python 3.9 provides `zoneinfo` as part of the standard libary.
```
>>> from zoneinfo import ZoneInfo
>>> datetime(2020, 10, 7, tzinfo=timezone("America/Los_Angeles")) \
  .isoformat()
'2020-10-07T00:00:00-07:00'
#                    ^^^^^
#                    ok!
```
- Had to remove mypy and black from requirements.txt due to
`typed-ast` not being having wheels for Python 3.9.
- Reorganize imports into groups.
- Migrations script, creates initial.